data_ingestion/mqtt_kafka_bridge/requirements.txt
# Copiato da data_ingestion/mqtt-kafka-bridge/requirements.txt
paho-mqtt
confluent-kafka
PyYAML 